+++ src/extension.ts
@@ -1,18 +1,25 @@
-import { ExtensionContext } from "vscode";
+import * as vscode from "vscode";
 import { registerCommands } from "./commands";
 import { registerCommentController } from "./comments";
 import { registerFileSystemProvider } from "./fileSystem";
+import { log } from "./logger";
 import { store } from "./store";
 import { initializeAuth } from "./store/auth";
 import { initializeStorage } from "./store/storage";
 import { registerTreeProvider } from "./tree";
 
-export function activate(context: ExtensionContext) {
-  registerCommands(context);
-  registerFileSystemProvider(store);
-  registerTreeProvider(store, context.extensionPath);
-  registerCommentController();
+export async function activate(context: vscode.ExtensionContext) {
+  try {
+    log.setLoggingChannel(vscode.window.createOutputChannel('GistPad'));
+    initializeAuth();
 
-  initializeStorage(context);
-  initializeAuth();
+    registerCommands(context);
+    registerFileSystemProvider(store);
+    registerTreeProvider(store, context.extensionPath);
+    registerCommentController();
+    initializeStorage(context);
+  } catch (e) {
+    log.error(e);
+    vscode.window.showErrorMessage(e);
+  }
 }