+++ src/interfaces/IClipboardToImageBuffer.ts
@@ -1,0 +1,5 @@
+
+// abstraction over the clipboard
+export interface IClipboardToImageBuffer {
+    getImageBits(): Promise<Buffer>;
+}