+++ src/utils/randomInt.ts
@@ -1,0 +1,3 @@
+export const randomInt = (base = 1000000) => {
+    return Math.round(Math.random() * base);
+}