+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## v0.0.14 (12/07/2019)
+
+- 🖼️ Implement the `Paste Screenshot` command that allows to paste a screenshot from your clipboard into the gist.
+
+  ![paste-screenshot](https://user-images.githubusercontent.com/1478800/70382701-9a7ac980-1914-11ea-9fb0-6e55424e2e54.gif)
+
 ## v0.0.13 (12/07/2019)
 
 - Gists are now sorted by their last update time, not alphabetically (in both the tree view and command palette)